+++ NEWS.md
@@ -8,3 +8,5 @@
 - Using solys v0.2.1a1.
 - Cross/Mesh informs the user of next step and number of remaining steps.
 - Removed minimum size, now it's just initial size.
+
+- Added missing imports