+++ NEWS.md
@@ -12,3 +12,5 @@
 - Added missing imports
 
 - Changed Set Position Azimuth maximum and minimum.
+- Set position input now has the current value initially.
+- Save log folder fixed.