+++ NEWS.md
@@ -16,3 +16,7 @@
 - Save log folder fixed.
 
 - Fixed bug with position widget
+
+- Adjust step set to 0.01.
+- Cross/Mesh now sais performed steps instead of remaining steps.
+- Current adjustment label rounded to fourth decimal.